+++ express/styles/styles.css
@@ -1,5 +1,12 @@
 /* default content */
 
+@font-face {
+  font-family: 'Trebuchet MS';
+  size-adjust: 90%;
+  src: local('Trebuchet MS'),
+       local('TrebuchetMS');
+}
+
 :root {
   --color-white: #FFF;
   --color-gray-100: #F8F8F8;
@@ -31,7 +38,7 @@
   /* body */
   --body-background-color: var(--color-white);
   --body-alt-background-color: var(--color-gray-200);
-  --body-font-family: 'adobe-clean', 'Adobe Clean', sans-serif;
+  --body-font-family: 'adobe-clean', 'Adobe Clean', 'Trebuchet MS', sans-serif;
   --body-serif-font-family: 'adobe-clean-serif', 'Adobe Clean Serif', serif;
   --body-font-weight: normal;
   --body-color: var(--color-gray-800);
@@ -362,8 +369,6 @@ main h1, h2, h3, h4, h5, h6 {
 
 main h1 {
   font-size: var(--heading-font-size-xxl);
-  opacity: 0;
-  transition: opacity 0.2s;
 }
 
 main h2 {
@@ -452,10 +457,6 @@ main .section-wrapper:last-of-type {
   main .section-wrapper h1 {
     font-size: var(--heading-font-size-l);
   }
-}
-
-.font-loaded main h1 {
-  opacity: 1;
 }
 
 /* hero */
